requirements.txt
streamlit
python-dotenv
langchain
db-sqlite3
langchain-community
langchain-groq